handle 0 shards

+++ Deployment/Containerized/usr/bin/philbot_containerized_auto_scale_sharding.sh
@@ -8,9 +8,11 @@ if [ "$SHARD_COUNT" -eq "$SHARD_COUNT_NEW" ]; then
   exit 0
 fi
 
-shard_container_names=$(echo discordgateway2http_$(seq -s " discordgateway2http_" 0 $(($SHARD_COUNT-1))))
-docker stop $shard_container_names
-docker rm $shard_container_names
+if [ "$SHARD_COUNT" -gt "0" ]; then
+  shard_container_names=$(echo discordgateway2http_$(seq -s " discordgateway2http_" 0 $(($SHARD_COUNT-1))))
+  docker stop $shard_container_names
+  docker rm $shard_container_names
+fi
 
 echo "SHARD_COUNT=$SHARD_COUNT_NEW" > /opt/philbot/shards
 bash /usr/bin/philbot_containerized_start.sh
